Trainer update
Now the trainer is able to read data from MNEST training dataset, and it
does train and return test results from the file. However, I have not
yet configured the parameters for training the network, so I haven't
gotten a well-trained network yet.
Additionally, it does not have functionality to export neural network
data, which is going to be something I'll be working on very soon.